requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
gradio
controlnet-aux
mediapipe
opencv-python
pillow
numpy
